+++ internal/ui/plugins/loader.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/devnullvoid/pvetui/internal/config"
 	"github.com/devnullvoid/pvetui/internal/ui/components"
+	"github.com/devnullvoid/pvetui/internal/ui/plugins/ansible"
 	"github.com/devnullvoid/pvetui/internal/ui/plugins/commandrunner"
 	"github.com/devnullvoid/pvetui/internal/ui/plugins/communityscripts"
 	"github.com/devnullvoid/pvetui/internal/ui/plugins/guestlist"
@@ -13,6 +14,7 @@ import (
 type factory func() components.Plugin
 
 var registry = map[string]factory{
+	ansible.PluginID:          func() components.Plugin { return ansible.New() },
 	commandrunner.PluginID:    func() components.Plugin { return commandrunner.New() },
 	communityscripts.PluginID: func() components.Plugin { return communityscripts.New() },
 	guestlist.PluginID:        func() components.Plugin { return guestlist.New() },